pass kubeconfig to controllers when running out-of-cluster
When we are running controllers locally against a remote cluster,
we need to be able to pass the kubeconfig file for that cluster.
This kubeconfig is then converted to a rest.Config and passed on
to the manager. Without the kubeconfig file, we will not be able
to run the controllers locally.

Signed-off-by: Umanga Chapagain <uchapaga@redhat.com>